+++ spec/spec_helper.rb
@@ -12,6 +12,7 @@ RSpec.configure do |config|
   config.color_enabled = true
 
   config.filter_run :focus => true
+  config.treat_symbols_as_metadata_keys_with_true_values = true
   config.run_all_when_everything_filtered = true
 
   config.before(:each) do